+++ lib/app/global/widgets/app_bar/components/components_app_bar.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:pscomidas/app/global/models/enums/filter.dart';
-import 'package:pscomidas/app/modules/home/home_page.dart';
 import 'package:pscomidas/app/modules/home/schemas.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter/material.dart';
@@ -20,8 +19,7 @@ class LogoAppBar extends StatelessWidget {
           "assets/images/logo.png",
           width: screen.width * 0.2,
         ),
-        onTap: () => Navigator.push(
-            context, MaterialPageRoute(builder: (context) => const HomePage())),
+        onTap: () => Navigator.popUntil(context, ModalRoute.withName('/')),
       ),
     );
   }
